requirements.txt
selenium
discord.py[voice]
python-dotenv
youtube-dl
python-opensubtitles
validators